fix email local storage
